Break out the extract accessibility command as a separate file
